Refactor search components
