+++ src/components/Search/OpenSearch/NFTsSearch/index.tsx
@@ -1,0 +1,34 @@
+import { IStep, SearchDefault } from 'components/Search/New'
+import { useEffect, useMemo, useState } from 'react'
+import { useFetchNFTsList } from 'utils/categories/nfts'
+
+interface INFTSearchProps {
+  step: IStep
+  preLoadedSearch: Array<{
+    name: string
+    route: string
+    logo: string
+  }>
+}
+
+export default function NFTsSearch({ step, preLoadedSearch }: INFTSearchProps) {
+  const [searchValue, setSearchValue] = useState('')
+  const [usePreloadedList, setUsePreloadedList] = useState(false)
+  const { data, loading } = useFetchNFTsList(searchValue)
+
+  useEffect(() => {
+    if (preLoadedSearch && !searchValue && !loading) setUsePreloadedList(true)
+    else setUsePreloadedList(false)
+  }, [preLoadedSearch, searchValue, loading])
+
+  const searchData = useMemo(() => {
+    const set = usePreloadedList ? preLoadedSearch : data ?? []
+    return set.map((el) => ({
+      name: el.name,
+      route: `/nfts/collection/${el.slug}`,
+      logo: el.logo,
+    }))
+  }, [data, usePreloadedList])
+
+  return <SearchDefault data={searchData} loading={loading} step={step} onSearchValueChange={setSearchValue} />
+}
